stripheaders v1.0
"stripheaders" is a utility for removing file headers from binary
VNMR data files, such that third party software does not need to deal
with the header files. The "pure data" information is stored in a new
file *.bin, the header information is collected in a separate file
*.hdr. In addition to that, "stripheaders" creates a text file (*.txt)
with readable information about the file structure. "stripheaders"
works for VNMR FIDs of any kind, for datdir/data (2D data only when
processed with "wft", NOT wft1d/wft2d), and datdir/phasefile (both
1D and 2D).